Crate me507_lab_3

Input capture of a PWM positive duty width to H-bridge motor driver

Read two channels of a wireless remote that outputs positive pulse widths of 1ms to 2ms

Control two motors (H-bridge) with the remote signals

Structs

Irqs 🔒

Statics

ALLOCATOR 🔒

Global allocator for the heap

ARENA 🔒

Size of the heap in bytes

DURATION_A 🔒

Positive pulse duration for channel A of remote in microseconds

DURATION_B 🔒

Positive pulse duration for channel B of remote in microseconds

DUTY_A 🔒

Percentage duty cycle for motor A, -100 to +100

DUTY_B 🔒

Percentage duty cycle for motor B, -100 to +100

LED_SPEED_SHARE 🔒

Functions

USART1 🔒 ⚠

__cortex_m_rt_main 🔒

__embassy_main 🔒

__make_static 🔒 ⚠

blinky 🔒

Task to blink the on-board LED at a speed proportional to motor speed

disable_motor 🔒

Disable a motor by disabling the PWM channels

enable_motor 🔒

Enable a motor by enabling the PWM channels

print_durations 🔒

Output the durations of the remote signals to the console

remote_ch_1 🔒

Task to read the remote channel 1 signal

remote_ch_2 🔒

Task to read the remote channel 2 signal

set_motor_duty 🔒

Set the duty cycle of a motor